Improved error messages
Give warnings or fatal errors when various things are amiss with the :cljsbuild
config. Print things to stderr or stdout, depending on what seems most
appropriate. Introduce a --debug flag.

Deal correctly with builds either specified as a map or a vector (cljsbuild
allows both). Fall back to the first build if no "dev" build is found.